requirements.txt
apiclient==1.0.2
bleach==1.4.2
certifi==2017.11.5
chardet==3.0.4
Django==2.0
django-common-helpers==0.7.0
django-debug-toolbar==1.4
django-filter==1.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.1
django-wysiwyg==0.7.1
djangorestframework==3.7.7
google-api-python-client==1.4.2
html5lib==0.9999999
httplib2==0.9.2
idna==2.6
lorem-ipsum-generator==0.3
Markdown==2.6.10
oauth2client==1.5.2
olefile==0.44
Pillow==4.3.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pytz==2017.3
requests==2.9.1
rsa==3.3
simplejson==3.8.2
six==1.10.0
sqlparse==0.1.18
transliterate==1.7.6
uritemplate==0.6
urllib3==1.22
